OSIF: resolve build-error
NA

+++ bee/rtl87x2g/osif/zephyr/osif_zephyr.c
@@ -153,29 +153,29 @@ bool os_alloc_secure_ctx_zephyr(uint32_t stack_size)
     return true;
 }
 
-K_THREAD_STACK_DEFINE(lowstack_stack, 768*4);
+K_THREAD_STACK_DEFINE(lowstack_stack, 768 * 4);
 struct k_thread lowstack_thread_data;
 bool os_task_create_zephyr(void **pp_handle, const char *p_name, void (*p_routine)(void *),
                            void *p_param, uint16_t stack_size, uint16_t priority)
 {
     k_tid_t ret_thread_handle;
 
-/* The lowstack thread must be able to preempt all native cooperative threads in Zephyr
-(such as bt tx thread, mesh adv thread, BT Mesh settings workq and so on..). 
-Otherwise, if lowstack thread is not processed in a timely manner within approximately 40ms, 
-it may result in lowstack crashing and subsequently causing no event reply after hci command is sent, leading to a hardfault. 
+    /* The lowstack thread must be able to preempt all native cooperative threads in Zephyr
+    (such as bt tx thread, mesh adv thread, BT Mesh settings workq and so on..).
+    Otherwise, if lowstack thread is not processed in a timely manner within approximately 40ms,
+    it may result in lowstack crashing and subsequently causing no event reply after hci command is sent, leading to a hardfault.
 
-Therefore, we need to make the lowstack thread a "Meta-IRQ thread", which can preempt coop threads and meta-irq threads.
-Specifically, the priority of lowstack thread is set to K_HIGHEST_THREAD_PRIO with CONFIG_NUM_METAIRQ_PRIORITIES=1.
+    Therefore, we need to make the lowstack thread a "Meta-IRQ thread", which can preempt coop threads and meta-irq threads.
+    Specifically, the priority of lowstack thread is set to K_HIGHEST_THREAD_PRIO with CONFIG_NUM_METAIRQ_PRIORITIES=1.
 
-Reference: https://docs.zephyrproject.org/latest/kernel/services/threads/index.html#meta-irq-priorities  */
+    Reference: https://docs.zephyrproject.org/latest/kernel/services/threads/index.html#meta-irq-priorities  */
 
     if (strcmp(p_name, "low_stack_task") == 0)
     {
         *pp_handle = &lowstack_thread_data;
-        ret_thread_handle = k_thread_create(&lowstack_thread_data, lowstack_stack, 768*4,
-                                (k_thread_entry_t) p_routine, p_param, NULL, NULL,
-                                K_HIGHEST_THREAD_PRIO, 0, K_MSEC(10));
+        ret_thread_handle = k_thread_create(&lowstack_thread_data, lowstack_stack, 768 * 4,
+                                            (k_thread_entry_t) p_routine, p_param, NULL, NULL,
+                                            K_HIGHEST_THREAD_PRIO, 0, K_MSEC(10));
         k_thread_name_set(ret_thread_handle, p_name);
         return true;
     }
@@ -1085,7 +1085,7 @@ bool os_timer_get_auto_reload_zephyr(void **pp_handle, long *p_autoreload)
 
         obj = (struct k_timer *)*pp_handle;
 
-        if(K_TIMEOUT_EQ(obj->period, K_NO_WAIT))
+        if (K_TIMEOUT_EQ(obj->period, K_NO_WAIT))
         {
             *p_autoreload = 0;
         }
@@ -1147,7 +1147,7 @@ extern void z_thread_timeout(struct _timeout *t);
 extern void sys_clock_announce_only_add_ticks(int32_t ticks);
 
 void os_pm_return_to_idle_task_zephyr(void)
-{ 
+{
     arch_kernel_init();//perform arm v81mainline initialization: including fault exception init & msp setting.
 
     RamVectorTableUpdate(GDMA0_Channel9_VECTORn, _isr_wrapper);
@@ -1217,11 +1217,11 @@ uint32_t os_sys_tick_clk_get_zephyr(void)
 
 uint32_t os_pm_next_timeout_value_get_zephyr(void)
 {
-/* Solution without the exclude timer mechanism. */
+    /* Solution without the exclude timer mechanism. */
     // uint32_t ticks = z_get_next_timeout_expiry();
     // return ticks;
 
-/* Solution with the exclude timer mechanism. */
+    /* Solution with the exclude timer mechanism. */
     uint32_t timeout_tick_res = 0xFFFFFFFF;
     uint32_t timeout_tick = 0;
     struct k_timer *timer;
@@ -1237,7 +1237,7 @@ uint32_t os_pm_next_timeout_value_get_zephyr(void)
         if (t->fn == z_timer_expiration_handler) //z_timer_expiration_handler
         {
             timer = CONTAINER_OF(t, struct k_timer, timeout);
-            T_OS_QUEUE_ELEM *p_cur_queue_item =lpm_excluded_handle[0].p_first;
+            T_OS_QUEUE_ELEM *p_cur_queue_item = lpm_excluded_handle[0].p_first;
             while (p_cur_queue_item != NULL)
             {
                 void *cur_excluded_handle = *(((PlatformPMExcludedHandleQueueElem *)p_cur_queue_item)->handle);
@@ -1250,7 +1250,7 @@ uint32_t os_pm_next_timeout_value_get_zephyr(void)
                         if (is_auto_reload)
                         {
                             DBG_DIRECT("[PM]!!handle=0x%x, exclude timer cannot be auto_reload", cur_excluded_handle);
-                            __ASSERT(0, "[PM]!!handle=0x%x", cur_excluded_handle);
+                            __ASSERT(0, "[PM]!!handle=0x%x", (unsigned int) cur_excluded_handle);
                         }
                         handle_checked = false;
                         break;
@@ -1307,7 +1307,8 @@ void os_pm_init_zephyr(void)
 
     platform_pm_register_callback_func_with_priority((void *)os_pm_check, PLATFORM_PM_CHECK, 1);
     platform_pm_register_callback_func_with_priority((void *)os_pm_store_zephyr, PLATFORM_PM_STORE, 1);
-    platform_pm_register_callback_func_with_priority((void *)os_pm_restore_zephyr, PLATFORM_PM_RESTORE, 1);
+    platform_pm_register_callback_func_with_priority((void *)os_pm_restore_zephyr, PLATFORM_PM_RESTORE,
+                                                     1);
     return;
 }
